precompile: move test helpers to their own file
